pyproject.toml
[project]
name = "mssql_mcp_server"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server that enables secure interaction with Microsoft SQL Server databases."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pymssql>=2.2.8",
    "pyodbc>=4.0.35",
    "azure-identity>=1.12.0",
    "azure-core>=1.24.0",
]

[tool.mcp]
system_dependencies.darwin = ["freetds"]
system_dependencies.linux = ["freetds-dev"]
system_dependencies.win32 = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mssql_mcp_server = "mssql_mcp_server:main"
